+++ crates/binding/src/js_plugin.rs
@@ -1,7 +1,7 @@
 use std::path::PathBuf;
 use std::sync::Arc;
 
-use crate::js_hook::{LoadResult, ResolveIdResult, TsFnHooks, WriteFile};
+use crate::js_hook::{LoadResult, ResolveIdParams, ResolveIdResult, TsFnHooks, WriteFile};
 
 pub struct JsPlugin {
     pub hooks: TsFnHooks,
@@ -9,7 +9,7 @@ pub struct JsPlugin {
 use anyhow::{anyhow, Result};
 use mako::ast::file::{Content, JsContent};
 use mako::compiler::Context;
-use mako::plugin::{Plugin, PluginGenerateEndParams, PluginLoadParam};
+use mako::plugin::{Plugin, PluginGenerateEndParams, PluginLoadParam, PluginResolveIdParams};
 use mako::resolve::{ExternalResource, Resolution, ResolvedResource, ResolverResource};
 
 impl Plugin for JsPlugin {
@@ -26,6 +26,17 @@ impl Plugin for JsPlugin {
 
     fn load(&self, param: &PluginLoadParam, _context: &Arc<Context>) -> Result<Option<Content>> {
         if let Some(hook) = &self.hooks.load {
+            if self.hooks.load_include.is_some()
+                && self
+                    .hooks
+                    .load_include
+                    .as_ref()
+                    .unwrap()
+                    .call(param.file.path.to_string_lossy().to_string())?
+                    == Some(false)
+            {
+                return Ok(None);
+            }
             let x: Option<LoadResult> = hook.call(param.file.path.to_string_lossy().to_string())?;
             if let Some(x) = x {
                 match x.content_type.as_str() {
@@ -53,11 +64,17 @@ impl Plugin for JsPlugin {
         &self,
         source: &str,
         importer: &str,
+        params: &PluginResolveIdParams,
         _context: &Arc<Context>,
     ) -> Result<Option<ResolverResource>> {
         if let Some(hook) = &self.hooks.resolve_id {
-            let x: Option<ResolveIdResult> =
-                hook.call((source.to_string(), importer.to_string()))?;
+            let x: Option<ResolveIdResult> = hook.call((
+                source.to_string(),
+                importer.to_string(),
+                ResolveIdParams {
+                    is_entry: params.is_entry,
+                },
+            ))?;
             if let Some(x) = x {
                 if let Some(true) = x.external {
                     return Ok(Some(ResolverResource::External(ExternalResource {