+++ crates/mako/src/generate/transform.rs
@@ -227,7 +227,6 @@ pub fn transform_js_generate(transform_js_param: TransformJsParam) -> Result<()>
                             to_replace: dep_map,
                             context,
                             unresolved_mark,
-                            top_level_mark,
                         };
                         ast.ast.visit_mut_with(&mut dep_replacer);
 
@@ -337,7 +336,7 @@ mod tests {
     }
 
     fn transform_css_code(content: &str, path: Option<&str>) -> String {
-        let path = if let Some(p) = path { p } else { "test.css" };
+        let path = path.unwrap_or("test.css");
         let context: Arc<Context> = Arc::new(Default::default());
         let mut ast = CssAst::build(path, content, context.clone(), false).unwrap();
         transform_css_generate(&mut ast.ast, &context);