+++ crates/mako/src/visitors/dynamic_import.rs
@@ -11,6 +11,7 @@ use swc_core::ecma::visit::{VisitMut, VisitMutWith};
 
 use super::dep_replacer::miss_throw_stmt;
 use crate::ast::utils::{is_dynamic_import, promise_all, require_ensure};
+use crate::ast::DUMMY_CTXT;
 use crate::compiler::Context;
 use crate::generate::chunk::ChunkId;
 use crate::module::generate_module_id;
@@ -44,11 +45,7 @@ impl<'a> VisitMut for DynamicImport<'a> {
             let insert_at = n
                 .body
                 .iter()
-                .position(|module_item| {
-                    !module_item
-                        .as_stmt()
-                        .map_or(false, |stmt| stmt.is_directive())
-                })
+                .position(|module_item| !module_item.directive_continue())
                 .unwrap();
 
             let (id, _) = self
@@ -86,8 +83,7 @@ impl<'a> VisitMut for DynamicImport<'a> {
                     if self
                         .dep_to_replace
                         .missing
-                        .get(source.value.as_ref())
-                        .is_some()
+                        .contains_key(source.value.as_ref())
                     {
                         call_expr.args[0] = ExprOrSpread {
                             spread: None,
@@ -161,18 +157,19 @@ impl<'a> VisitMut for DynamicImport<'a> {
 
                         let require_module = generate_module_id(&resolved_info.1, &self.context);
 
-                        let lazy_require_call = member_expr!(DUMMY_SP, __mako_require__.bind)
-                            .as_call(
+                        let lazy_require_call =
+                            member_expr!(DUMMY_CTXT, DUMMY_SP, __mako_require__.bind).as_call(
                                 DUMMY_SP,
                                 vec![
                                     quote_ident!("__mako_require__").as_arg(),
                                     quote_str!(require_module.as_str()).as_arg(),
                                 ],
                             );
-                        let dr_call = member_expr!(DUMMY_SP, __mako_require__.dr).as_call(
-                            DUMMY_SP,
-                            vec![self.interop.clone().as_arg(), lazy_require_call.as_arg()],
-                        );
+                        let dr_call = member_expr!(DUMMY_CTXT, DUMMY_SP, __mako_require__.dr)
+                            .as_call(
+                                DUMMY_SP,
+                                vec![self.interop.clone().as_arg(), lazy_require_call.as_arg()],
+                            );
 
                         member_expr!(@EXT, DUMMY_SP, load_promise.into(), then)
                             .as_call(call_expr.span, vec![dr_call.as_arg()])