+++ e2e/fixtures/plugins/plugins.config.js
@@ -11,18 +11,19 @@ module.exports = [
     }
   },
   {
-    async load(path) {
-      if (path.endsWith('.bar')) {
-        return {
-          content: `export default () => <Foooo>.bar</Foooo>;`,
-          type: 'jsx',
-        };
-      }
+    async loadInclude(path) {
+      return path.endsWith('.bar');
+    },
+    async load() {
+      return {
+        content: `export default () => <Foooo>.bar</Foooo>;`,
+        type: 'jsx',
+      };
     }
   },
   {
-    async resolveId(source, importer) {
-      console.log('resolveId', source, importer);
+    async resolveId(source, importer, options) {
+      console.log('resolveId', source, importer, options);
       if (source === 'resolve_id') {
         return { id: require('path').join(__dirname, 'resolve_id_mock.js'), external: false };
       }